Move abstract type data structures into their own modules
